+++ android/app/google-services.json
@@ -29,6 +29,13 @@
             {
               "client_id": "249746516769-btch40eqce6b1v4mtmel74pmsi2fm2np.apps.googleusercontent.com",
               "client_type": 3
+            },
+            {
+              "client_id": "249746516769-4fa86l4h0l7meabvkb704uaoq7ngvuav.apps.googleusercontent.com",
+              "client_type": 2,
+              "ios_info": {
+                "bundle_id": "dev.onenineone.cashPlaner"
+              }
             }
           ]
         }
