+++ lib/app/core/enums.dart
@@ -1,0 +1,6 @@
+enum Status {
+  loading,
+  initial,
+  error,
+  success,
+}
