+++ lib/domain/models/expense_model.dart
