+++ lib/features/auth/cubit/auth_state.dart
@@ -1,0 +1,13 @@
+part of 'auth_cubit.dart';
+
+class AuthState {
+  final UserModel? user;
+  final Status status;
+  final String erroeMessage;
+
+  AuthState({
+    this.user,
+    this.status = Status.initial,
+    this.erroeMessage = '',
+  });
+}
