+++ pubspec.yaml
@@ -46,6 +46,10 @@ dependencies:
   image_cropper: ^3.0.0
   intl: ^0.17.0
   go_router: ^4.3.0
+  flutterfire_ui: ^0.4.3+14
+  google_sign_in: ^5.4.2
+  freezed_annotation: ^2.2.0
+  json_annotation: ^4.7.0
 
 dev_dependencies:
   flutter_test:
@@ -57,6 +61,9 @@ dev_dependencies:
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
   flutter_lints: ^2.0.0
+  build_runner: ^2.3.0
+  freezed: ^2.2.0
+  json_serializable: ^6.5.1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
